+++ src/main/kotlin/com/github/xepozz/git_churn/config/StringColumn.kt
@@ -1,0 +1,5 @@
+package com.github.xepozz.git_churn.config
+
+data class StringColumn(
+    var value: String,
+)