+++ src/main/kotlin/com/github/xepozz/git_churn/listener/StartupListener.kt
@@ -1,11 +1,10 @@
-package com.github.xepozz.git_churn
+package com.github.xepozz.git_churn.listener
 
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.startup.ProjectActivity
 
 class StartupListener() : ProjectActivity {
     override suspend fun execute(project: Project) {
-        val service = project.getService(GitChurnService::class.java)
-        service.refresh()
+        GitChurnListener.fireSettingsUpdated()
     }
 }